requirements.txt
transformers
torch
python-dotenv
streamlit
fastapi
uvicorn
